requirements.txt
# UI web
streamlit>=1.32.0,<2.0.0

# Cliente LLM
openai>=1.0.0,<2.0.0
google-cloud-aiplatform>=1.65.0,<2.0.0

# Variables de entorno
python-dotenv>=1.0.0,<2.0.0

# Exportación PDF
reportlab>=4.0.0,<5.0.0

# Testing
pytest>=7.0.0,<9.0.0